requirements.txt
Flask
Flask-CORS
pyserial
opencv-python
numpy
Pillow

thefuzz
python-Levenshtein
dateparser
python-dateutil